Updated schema regarding organization
I have made a post-demo change to the domain model. There was one new property added: has:Job to relate an orgabization with a job. The only other changes were to the domains of hasCompetencyProfile, hasJobDescription & hasJobPosting that now point from the Job (i.e., the old old JobMaster) instead of from Organization.